+++ src/db.json
@@ -1,0 +1,33 @@
+
+[
+    {
+        "id": 1, 
+        "itemName": "aerosol cans",
+        "itemImg": "",
+        "actionList": [1, 2, 3]
+    },
+    {
+        "id": 2,
+        "itemName": "adhesive strips (band aids)", 
+        "itemImg": "",
+        "actionList": [1, 2]
+    },
+    {
+        "id": 3,
+        "itemName": "aluminium cans", 
+        "itemImg": "",
+        "actionList": [2, 3]
+    },
+    {
+        "id": 4,
+        "itemName": "alumninuim foil", 
+        "itemImg": "",
+        "actionList": [1, 3]
+    },
+    {
+        "id": 5,
+        "itemName": "aluminium trays", 
+        "itemImg": "",
+        "actionList": [2]
+    }
+]
